Moved scrolltotop, update and submit form components
